+++ Assets/Authoring/Venue/Unique/VenueProfile.asset
@@ -114,6 +114,22 @@ MonoBehaviour:
   dirtIntensity:
     m_OverrideState: 0
     m_Value: 0
+--- !u!114 &-7311563954459877882
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e6736c5e0a4349fcbfb2ab3f96e0389c, type: 3}
+  m_Name: PosterizeComponent
+  m_EditorClassIdentifier: 
+  active: 1
+  Steps:
+    m_OverrideState: 0
+    m_Value: 0
 --- !u!114 &-7172459413072436801
 MonoBehaviour:
   m_ObjectHideFlags: 3
@@ -412,15 +428,31 @@ MonoBehaviour:
   - {fileID: -8201876109855894492}
   - {fileID: 7766579603084563493}
   - {fileID: 219214101893045267}
-  - {fileID: 4615704651314425003}
-  - {fileID: -8053512828986670601}
-  - {fileID: -6923786053771193734}
   - {fileID: -6081049760158747905}
   - {fileID: -6896316964874578581}
   - {fileID: -7172459413072436801}
-  - {fileID: -9023157280191997172}
-  - {fileID: -8639306260386040192}
   - {fileID: -92697382740459818}
+  - {fileID: 1446895283939926279}
+  - {fileID: 8105415359043762657}
+  - {fileID: -7311563954459877882}
+  - {fileID: 5972827141023902039}
+  - {fileID: 186418053702648674}
+--- !u!114 &186418053702648674
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: afb7ebeff3a44c3aadeafb3364cec179, type: 3}
+  m_Name: TrailsComponent
+  m_EditorClassIdentifier: 
+  active: 1
+  length:
+    m_OverrideState: 0
+    m_Value: 0.5
 --- !u!114 &219214101893045267
 MonoBehaviour:
   m_ObjectHideFlags: 3
@@ -461,6 +493,31 @@ MonoBehaviour:
   blueOutBlueIn:
     m_OverrideState: 0
     m_Value: 10
+--- !u!114 &1446895283939926279
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 69419bebc4674f43ae70a773686ab6a9, type: 3}
+  m_Name: MirrorComponent
+  m_EditorClassIdentifier: 
+  active: 1
+  enabled:
+    m_OverrideState: 0
+    m_Value: 0
+  wipeIndex:
+    m_OverrideState: 0
+    m_Value: 3
+  wipeTime:
+    m_OverrideState: 0
+    m_Value: 0.5
+  startTime:
+    m_OverrideState: 0
+    m_Value: 0
 --- !u!114 &4615704651314425003
 MonoBehaviour:
   m_ObjectHideFlags: 3
@@ -480,6 +537,22 @@ MonoBehaviour:
   scanlineCount:
     m_OverrideState: 0
     m_Value: 270
+--- !u!114 &5972827141023902039
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 96d131c3af4147d9b3eb0bcaf017d0d3, type: 3}
+  m_Name: SlowFPSComponent
+  m_EditorClassIdentifier: 
+  active: 1
+  SkipFrames:
+    m_OverrideState: 0
+    m_Value: 1
 --- !u!114 &6786810113853146721
 MonoBehaviour:
   m_ObjectHideFlags: 3
@@ -552,3 +625,22 @@ MonoBehaviour:
   saturation:
     m_OverrideState: 0
     m_Value: -0
+--- !u!114 &8105415359043762657
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 676f25d64ce84c15a1041f1bfaafbf15, type: 3}
+  m_Name: ScanlineComponent
+  m_EditorClassIdentifier: 
+  active: 1
+  intensity:
+    m_OverrideState: 0
+    m_Value: 0
+  scanlineCount:
+    m_OverrideState: 0
+    m_Value: 270
