--- Assets/Script/Venue/VenueCamera/MirrorComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: bd8bfb7e8840474ab6d9dc0b57f30be5
-timeCreated: 1752112569
