--- Assets/Script/Venue/VenueCamera/PosterizeComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3e96a48f57254225851f99948b446de5
-timeCreated: 1752614418
