--- Assets/Script/Venue/VenueCamera/ScanlineComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3289a0ccc8054c71b8f0411d573ed049
-timeCreated: 1752101277
