--- Assets/Script/Venue/VenueCamera/SlowFPSComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1a5a76be172c43bd98aa6b69c6dabbeb
-timeCreated: 1752636509
