--- Assets/Script/Venue/VenueCamera/TrailsComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 07310bc8d8814c35a7ee2f0c0261441a
-timeCreated: 1752190884
