--- Assets/Script/Venue/VenueCamera/VenuePostProcessPass.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: baf20299ef8d4ce395c5ad8adbd8242e
-timeCreated: 1752105333
