+++ Assets/Script/Venue/VolumeComponents/MirrorComponent.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Runtime.CompilerServices;
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 
-namespace YARG.Venue.VenueCamera
+namespace YARG.Venue.VolumeComponents
 {
     [Serializable]
     [VolumeComponentMenuForRenderPipeline("Venue/Mirror", typeof(UniversalRenderPipeline))]
