+++ Assets/Script/Venue/VolumeComponents/PosterizeComponent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e6736c5e0a4349fcbfb2ab3f96e0389c
+timeCreated: 1754082260
