+++ Assets/Script/Venue/VolumeComponents/ScanlineComponent.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 
-namespace YARG.Venue.VenueCamera
+namespace YARG.Venue.VolumeComponents
 {
     [Serializable]
     [VolumeComponentMenuForRenderPipeline("Venue/Scanline", typeof(UniversalRenderPipeline))]
