+++ Assets/Script/Venue/VolumeComponents/TrailsComponent.cs
@@ -2,7 +2,7 @@
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 
-namespace YARG.Venue.VenueCamera
+namespace YARG.Venue.VolumeComponents
 {
     [Serializable]
     [VolumeComponentMenuForRenderPipeline("Venue/Trails", typeof(UniversalRenderPipeline))]
